fix: correctness, data loss, and verification that never ran (phase 1)
Sphere theta cap: for rmin=0 the theta surface is the cone swept by
r in [0, rmax], whose apex is the origin. The mesher placed the apex at
(0, 0, rmax*cos(theta)), making the solid a spherical zone rather than a
sector while still emitting cone normals. Measured error ran from -60.4%
at deltatheta=45deg to +12.5% at 120deg, vanishing only at 90deg where the
cut plane passes through the origin. The mesh stayed closed and correctly
wound, so assert_closed_outward passed on it -- only a volume check catches
this. Same origin-vs-axis confusion in add_phi_wedge_face. Three regression
tests added, verified to fail on the previous code with exactly those errors.

Material state/MEE: Material carries state and mee and the serializer emits
both, but MaterialInfo declared neither and the edit paths rebuilt the
payload field by field. update_material replaces wholesale, so typing one
character in Density erased state="gas" and <MEE> from the document and the
export. Reachable on sample_data/pod_asm.gdml. Fixed by typing the fields
and spreading the original; ElementInfo gains fractions to defuse the same
trap in the unused updateElement path.

Type checking: tsconfig.json is solution-style ("files": [] plus project
references), so the documented `tsc --noEmit` resolved zero input files and
exited 0 regardless of errors -- in the README and in both run scripts.
Switched to `tsc -b`. Enabling real checks surfaced 3 pre-existing ESLint
errors and 16 clippy warnings, all fixed; both now pass clean, clippy under
-D warnings.

CI: added, since none existed. cargo fmt --check is left commented out --
the tree has never been rustfmt-clean (113 hunks across 21 files) and the
gate would fail on arrival.

Docs: Save/Save As are browser downloads and never touch the opened file;
the README claimed the opposite and the in-app confirm warned about an
overwrite that cannot happen. Node version corrected to Vite 7's actual
requirement, sample tables regenerated from real sizes with provenance,
frontend/README.md replaced (was untouched Vite boilerplate).

+++ backend/src/mesh/primitives/cut_tube_mesh.rs
@@ -191,7 +191,7 @@ pub fn tessellate_cut_tube(
                 positions.push(z as f32);
                 normals.push(nx as f32);
                 normals.push(ny as f32);
-                normals.push(0.0 as f32);
+                normals.push(0.0_f32);
             }
 
             if flip {
